+++ Realm.Shared/Realm.cs
@@ -562,6 +562,11 @@ namespace Realms
             NativeTable.remove_row(tableHandle, (RowHandle)obj.RowHandle);
         }
 
+        /// <summary>
+        /// Remove objects matcing a query from the realm
+        /// </summary>
+        /// <typeparam name="T">Type of the objects to remove</typeparam>
+        /// <param name="range">The query to match for</param>
         public void RemoveRange<T>(RealmResults<T> range) where T: RealmObject
         {
             if (!IsInTransaction)
@@ -570,6 +575,10 @@ namespace Realms
             NativeResults.clear(range.ResultsHandle);
         }
 
+        /// <summary>
+        /// Remove all objects of a type from the realm
+        /// </summary>
+        /// <typeparam name="T">Type of the objects to remove</typeparam>
         public void RemoveAll<T>() where T: RealmObject
         {
             if (!IsInTransaction)